+++ gate/__init__.py
@@ -1,0 +1,3 @@
+from .safety_gate_controller import ESP32ControlNode
+
+__all__ = ['ESP32ControlNode']